Code refactoring
Code refactoring, exception handling, java docs

+++ java-challenge-master/src/main/java/jp/co/axa/apidemo/repositories/EmployeeRepository.java
@@ -4,6 +4,11 @@ import jp.co.axa.apidemo.entities.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Employee Repository interface to interact with database for performing CRUD operations
+ * @author Sumanjula Roy
+ * @version 1.0
+ */
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee,Long> {
+public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
 }
